Create new function f2bpurge to clean-out fail2ban logs
This function works in conjunction with f2bnet to administer fail2ban on
the computer.

The function f2bnet provides a list of the most-commonly banned /16
networks as found in the fail2ban log files.  When the number of times
that a given network is banned starts to grow, it makes sense to block
all traffic originating from the given network. (The number of times
banned before being added to the firewall rules is arbitrary and the
blocking at the firewall-level is done manually in anycase)

Once an IP network is added to the list of banned networks in the
firewall, it is no longer of importance to keep those IP addresses in
the fail2ban logs, since we expect no further traffic at all originating
from the network in question.  This function, therefore, facilitates
cleaning out the log files, eliminating all lines referrring to the
networks in question. This ensures that in the future, when f2bnet is
run, only newly active networks will be listed.

It is of importance to note that the bulk of the process is done
manually:
- viewing list of banned networks (with f2bnet)
- adding network to the firewall banned network rules

Once those tasks are done, this script can be run, which will use sed to
delete all lines referring to the 'n' most commonly occuring networks in
the fail2ban logs

This isn't fool-proof, and there are a number of things that should be
done to ensure that this administation task runs properly.

In a separate commit, need to:
- Backup/Archive the log files to a tar.xz file with today's date
- Gunzip any fail2ban log files which were gzipped by logrotate
- re-gzip the applicable files once purged of network references